removed geode-for-redis shutdown command
